+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/VisitorsWidget/useVisitors/parse.ts
@@ -1,13 +1,16 @@
 import moment, { Moment } from 'moment';
 
 import { VisitorsResponse } from 'api/graph_data_units/responseTypes';
+import {
+  TimeSeries,
+  TimeSeriesResponseRow,
+  TimeSeriesRow,
+} from 'api/graph_data_units/responseTypes/VisitorsWidget';
 
 import { timeSeriesParser } from 'components/admin/GraphCards/_utils/timeSeries';
 import { IResolution } from 'components/admin/ResolutionControl';
 
 import { get } from 'utils/helperUtils';
-
-import { TimeSeries, TimeSeriesResponseRow, TimeSeriesRow } from './typings';
 
 export const getEmptyRow = (date: Moment) => ({
   date: date.format('YYYY-MM-DD'),
